Segundo avance muestra las lista de los clientes por medio del store
